pantry_management_project/requirements.txt
streamlit
crewai
crewai-tools
python-dotenv
requests
pandas
numpy
PyYAML
jsonschema
aiohttp
openai
google-generativeai
boto3
s3fs
fastapi
uvicorn
pydantic
dataclasses-json
transformers
torch
tqdm
asyncio
jinja2
markdown-it-py
watchdog
altair
protobuf
scipy
matplotlib
seaborn
scikit-learn
fastavro
